requirements.txt
alembic==1.4.0
blinker==1.4
cffi==1.15.0
Click==7.0
cryptography==36.0.1
Flask==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.1
MarkupSafe==1.1.1
Pillow==8.4.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
WTForms==2.2.1
